Use explicit interoperability mechanism
The "HtmlParser2 family" of modules expects "node" objects to define a
non-standard data property named `tag`. Previously, this requirement was
informally circumvented through a brittle mechanism (patching internal
methods of various modules).

With the release of npm version 3, the location of the modules in the
file system is no longer stable, making the prior solution untenable.
(In addition, future versions of Node.js will likely disallow consumer
modification of external modules, further invalidating this solution.)

This library already defined a set of non-standard accessor properties
specifically to satisfy the expectations of the third party libraries in
use.

Define an additional non-standard accessor property on the Node
prototype in order to satisfy the requirements of those libraries.